Move conversation settings into config package
